Vu chatbot -NEW/requirements.txt
fastapi
uvicorn
chromadb
google-genai
python-dotenv
pandas
openpyxl
deep-translator
gTTS
SpeechRecognition
pydantic
